+++ balances-tx-payment/rpc/runtime-api/Cargo.toml
@@ -6,10 +6,10 @@ edition = "2021"
 
 [dependencies]
 # local deps
-encointer-balances-tx-payment = {path = "../..", default-features = false}
 encointer-primitives = { path = "../../../primitives", default-features = false }
 
 # substrate deps
+codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
 frame-support = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
 sp-api = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
 sp-std = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
